+++ HtmlRun.Runtime/Constants/NetworkInstructionsSet.cs
@@ -1,0 +1,6 @@
+namespace HtmlRun.Runtime.Constants;
+
+public static class NetworkInstructionsSet
+{
+  public const string HttpGet = "HttpGet";
+}